A separate file for BrickComputationGraph
